dev-4556 move area paths to component
