+++ src/js/components/award/shared/activity/ContractGrantActivityChartCircles.jsx
@@ -3,26 +3,55 @@
  * Created by Jonathan Hill 04/23/2020
  */
 
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
+import { formatMoney } from 'helpers/moneyFormatter';
 
 const propTypes = {
-    circles: PropTypes.arrayOf(PropTypes.shape({
-        key: PropTypes.string,
-        description: PropTypes.string,
-        className: PropTypes.string,
-        cx: PropTypes.number,
-        cy: PropTypes.number,
-        r: PropTypes.number
-    }))
+    transactions: PropTypes.array,
+    padding: PropTypes.object,
+    xScale: PropTypes.func,
+    yScale: PropTypes.func,
+    xAxisSpacing: PropTypes.number,
+    height: PropTypes.number
 };
 
-const ContractGrantActivityChartCircles = ({ circles }) => {
-    if (!circles.length) return null;
+const ContractGrantActivityChartCircles = ({
+    transactions,
+    padding,
+    xScale,
+    yScale,
+    xAxisSpacing,
+    height
+}) => {
+    // circle data
+    const [circleData, setCircleData] = useState([]);
+    // sets circle data - hook - on mount and when transactions change
+    useEffect(() => {
+        if (xScale && yScale) {
+            const circles = transactions
+                .map((data, i) => ({
+                    key: `${data.federal_action_obligation}${i}`,
+                    description: `A circle representing the transaction date, ${data.action_date.format('MM-DD-YYYY')} and running total obligation of ${formatMoney(data.running_obligation_total)}`,
+                    className: 'transaction-date-circle',
+                    cx: xScale(data.action_date.valueOf()) + padding.left,
+                    cy: (height - yScale(data.running_obligation_total)),
+                    r: 1.5
+                }));
+            setCircleData(circles);
+        }
+    }, [
+        transactions,
+        padding,
+        xScale,
+        yScale,
+        xAxisSpacing,
+        height
+    ]);
     return (
         <g className="contract-grant-activity-chart__circles">
             {
-                circles.map((circle) => {
+                circleData.map((circle) => {
                     const {
                         key,
                         description,
