+++ src/js/components/award/shared/activity/ContractGrantActivityChartVerticalLines.jsx
@@ -5,6 +5,7 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
+import moment from 'moment';
 import VerticalLine from 'components/sharedComponents/VerticalLine';
 
 const propTypes = {
@@ -34,6 +35,26 @@ const ContractGrantActivityChartVerticalLines = ({
     const endLineText = awardType === 'grant' ? 'End' : 'Current End';
     // class name for end line and text
     const endLineClassName = awardType === 'grant' ? 'grant-end' : 'contract-end';
+    const lineData = [
+        {
+            text: 'Start Date',
+            date: startLineValue
+        },
+        {
+            text: 'Todays Date',
+            date: todayLineValue
+        },
+        {
+            text: 'End Date',
+            date: endLineValue
+        },
+        {
+            text: 'Potential End Date',
+            date: potentialEndLineValue
+        }
+    ];
+    const descriptions = [startLineValue, todayLineValue, endLineValue, potentialEndLineValue]
+        .map((line, i) => `A vertical line representing the ${lineData[i].text}, ${moment(lineData[i].date).format("dddd, MMMM Do YYYY") || ''}`);
     return (
         <g className="contract-grant-activity-chart__vertical-lines">
             {/* start line */}
@@ -43,6 +64,7 @@ const ContractGrantActivityChartVerticalLines = ({
                 y2={height}
                 textY={0}
                 text="Start"
+                description={descriptions[0]}
                 xMax={xDomain[1]}
                 xMin={xDomain[0]}
                 xValue={startLineValue}
@@ -57,6 +79,7 @@ const ContractGrantActivityChartVerticalLines = ({
                 y2={height}
                 textY={0}
                 text="Today"
+                description={descriptions[1]}
                 xMax={xDomain[1]}
                 xMin={xDomain[0]}
                 xValue={todayLineValue}
@@ -71,6 +94,7 @@ const ContractGrantActivityChartVerticalLines = ({
                 y2={height}
                 textY={0}
                 text={endLineText}
+                description={descriptions[2]}
                 xMax={xDomain[1]}
                 xMin={xDomain[0]}
                 xValue={endLineValue}
@@ -85,6 +109,7 @@ const ContractGrantActivityChartVerticalLines = ({
                 y2={height}
                 textY={0}
                 text="Potential End"
+                description={descriptions[3]}
                 xMax={xDomain[1]}
                 xMin={xDomain[0]}
                 xValue={potentialEndLineValue}
